+++ CHANGELOG.md
@@ -17,7 +17,9 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 ## [2020.2.1.20] - 2020-04-20
 
 ### Added
-- added structs for order and user
+- added structs for user, order, operation and workplaces
+- fully working user input with checking in syteline
+- updated rostra.html file 
 
 
 ## [2020.2.1.16] - 2020-04-16